requirements.txt
absl-py==2.4.0
addict==2.4.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
-e file:///home/arota/spectra/mast3r/asmk
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
brotli==1.2.0
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
comm==0.2.3
configargparse==1.7.5
contourpy==1.3.3
cuda-bindings==13.0.3
cuda-pathfinder==1.2.2
cycler==0.12.1
cython==3.2.4
dash==4.0.0
decorator==5.2.1
einops==0.8.2
executing==2.2.1
faiss-gpu-cu12==1.14.1.post1
fastapi==0.135.1
fastjsonschema==2.21.2
ffmpy==1.0.0
filelock==3.20.0
flask==3.1.3
fonttools==4.62.1
fsspec==2025.12.0
gradio==6.9.0
gradio-client==2.3.0
groovy==0.1.2
grpcio==1.78.0
gsplat==1.5.3
h11==0.16.0
hf-xet==1.4.2
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.7.1
idna==3.11
importlib-metadata==8.7.1
ipython==9.10.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
itsdangerous==2.2.0
jaxtyping==0.3.9
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter-core==5.9.1
jupyterlab-widgets==3.0.16
kiwisolver==1.5.0
markdown==3.10.2
markdown-it-py==4.0.0
markupsafe==3.0.2
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
narwhals==2.18.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
ninja==1.13.0
numpy==2.4.3
nvidia-cublas==13.1.0.3
nvidia-cublas-cu12==12.9.1.4
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cuda-runtime-cu12==12.9.79
nvidia-cudnn-cu13==9.15.1.9
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
open3d==0.19.0
open3d-python==0.3.0.0
opencv-python==4.13.0.92
orjson==3.11.7
packaging==26.0
pandas==3.0.1
parso==0.8.6
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.9.4
plotly==6.6.0
prompt-toolkit==3.0.52
protobuf==7.34.0
ptyprocess==0.7.0
pure-eval==0.2.3
pyaml==26.2.1
pyarrow==23.0.1
pydantic==2.12.5
pydantic-core==2.41.5
pydub==0.25.1
pyglet==1.5.31
pygments==2.19.2
pyparsing==3.3.2
pyquaternion==0.9.9
python-dateutil==2.9.0.post0
python-multipart==0.0.22
pytz==2026.1.post1
pyyaml==6.0.3
referencing==0.37.0
requests==2.32.5
rerun-sdk==0.30.1
retrying==1.4.2
rich==14.3.3
roma==1.5.6
rpds-py==0.30.0
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
semantic-version==2.10.0
setuptools==82.0.1
shellingham==1.5.4
six==1.17.0
stack-data==0.6.3
starlette==0.52.1
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
textual==8.2.4
threadpoolctl==3.6.0
tomlkit==0.13.3
torch==2.10.0+cu130
torchvision==0.25.0+cu130
tqdm==4.67.3
traitlets==5.14.3
trimesh==4.11.3
triton==3.6.0
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.41.0
wadler-lindig==0.1.7
wcwidth==0.6.0
werkzeug==3.1.6
widgetsnbextension==4.0.15
zipp==3.23.0
